RR-09 Service implementation - [x] created UserService and PostService and Moved all the core logic from controllers to the services
