+++ app/Http/Controllers/PostController.php
@@ -2,32 +2,23 @@
 
 namespace App\Http\Controllers;
 
-use App\ApiToken;
-use App\Image;
-use App\Post;
-use App\Services\FileManager;
-use App\User;
+use App\Services\PostService;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Response;
+use Response;
 
 
 class PostController extends Controller
 {
-    protected $user;
-    protected $post;
     protected $response = [];
-    protected $fileManager;
-    protected $apiToken;
-    protected $image;
+    protected $postService;
 
-    public function __construct(Post $post, User $user, FileManager $fileManager, ApiToken $apiToken, Image $image)
+    /**
+     * PostController constructor.
+     * @param PostService $postService
+     */
+    public function __construct(PostService $postService)
     {
-        $this->fileManager = $fileManager;
-        $this->user        = $user;
-        $this->post        = $post;
-        $this->apiToken    = $apiToken;
-        $this->image       = $image;
+        $this->postService = $postService;
     }
 
     /**
@@ -37,37 +28,12 @@ class PostController extends Controller
      */
     public function savePost(Request $request)
     {
-        $data = $this->fetchDataFromRequest($request);
-        $post = $this->post->create($data);
-        if ($post) {
-            if ($request->hasFile('images')) {
-                $files = $request->file('images');
-                $post  = $this->fileManager->saveFile($post, $files, "post");
-            }
-            $this->response['code']    = "1000";
-            $this->response['message'] = "Post added successfully";
-            $this->response['post']    = $post;
-
-            return Response::json($this->response);
-        } else {
-            $this->response['code']    = "1001";
-            $this->response['message'] = "Problem adding a Post";
-
-            return Response::json($this->response);
-        }
-    }
-
-    /**
-     * returns all the posts
-     * @return mixed
-     */
-    public function fetchAllPost()
-    {
-        $posts                  = Post::All();
-        $this->response['post'] = $this->getPostDetails($posts);
-
+        $this->response = $this->postService->savePost($request);
         return Response::json($this->response);
+
     }
+
+
 
     /**
      * returns only the logged-in user's posts
@@ -76,16 +42,7 @@ class PostController extends Controller
      */
     public function fetchPersonalPost($apiToken)
     {
-        $userId = $this->getLoggedInUserId($apiToken);
-        $posts  = $this->post->where('userId', $userId)->get();
-        if ($posts) {
-            $this->response['code']    = "0000";
-            $this->response['message'] = "Posts fetched successfully";
-            $this->response['post']    = $this->getPostDetails($posts);
-        } else {
-            $this->response['code']    = '0001';
-            $this->response['message'] = 'No posts to display';
-        }
+      $this->response = $this->postService->fetchPersonalPost($apiToken);
 
         return Response::json($this->response);
     }
@@ -97,10 +54,7 @@ class PostController extends Controller
      */
     public function fetchPost($postType)
     {
-        $posts                     = $this->post->where('postType', $postType)->get();
-        $this->response['code']    = "0000";
-        $this->response['message'] = "Posts fetched successfully";
-        $this->response['post']    = $this->getPostDetails($posts);
+        $this->response = $this->postService->fetchPost($postType);
 
         return Response::json($this->response);
     }
@@ -112,73 +66,8 @@ class PostController extends Controller
      */
     public function fetchPostOfParticularArea(Request $request)
     {
-        $latitude                  = $request->latitude;
-        $longitude                 = $request->longitude;
-        $radius                    = $request->radius;
-        $posts                     = $this->post->where('postType', '=', $request->postType)
-            ->whereBetween('latitude', [$latitude - 0.018 * $radius, $latitude + 0.018 * $radius])
-            ->whereBetween('longitude', [$longitude - 0.018 * $radius, $longitude + 0.018 * $radius])->get();
-        $this->response['post']    = $this->getPostDetails($posts);
-        $this->response['code']    = "0000";
-        $this->response['message'] = "Posts fetched successfully";
+        $$this->response = $this->postService->fetchPostOfParticularArea($request);
 
         return Response::json($this->response);
-    }
-
-    /**
-     * get the details of jparticular type of posts
-     * @param $posts
-     * @return array
-     */
-    public function getPostDetails($posts)
-    {
-        $completePost = [];
-        foreach ($posts as $post) {
-            $filenameList = $this->image->where('postId', $post->id)->pluck('filename');
-            $images       = [];
-            foreach ($filenameList as $filename) {
-                array_push($images, route('file.get', $filename));
-            }
-            $post['images'] = $images;
-            array_push($completePost, $post);
-        }
-
-        return ($completePost);
-    }
-
-    /**
-     * returns the data from the request
-     * @param Request $request
-     * @return array
-     */
-    public function fetchDataFromRequest(Request $request)
-    {
-        $userId = $this->getLoggedInUserId($request->header('Authorization'));
-
-        return ([
-            'userId'        => $userId[0],
-            'location'      => $request->location,
-            'latitude'      => $request->latitude,
-            'longitude'     => $request->longitude,
-            'numberOfRooms' => $request->numberOfRooms,
-            'type'          => $request->type,
-            'description'   => $request->description,
-            'price'         => $request->price,
-            'postType'      => $request->postType,
-        ]);
-    }
-
-    /**
-     * returns the logged in user
-     * @param $header
-     * @return mixed
-     */
-    public function getLoggedInUserId($header)
-    {
-        $string   = expode(" ", $header);
-        $apiToken = $string[1];
-        $userId   = $this->apiToken->where('apiToken', $apiToken)->pluck('userId');
-
-        return $userId;
     }
 }
